refactor: modularize warehouse import dialog
